requirements.txt
flask==2.2.3
prometheus-client==0.16.0
psycopg2-binary==2.9.3
werkzeug==2.2.3
